refactor --

+++ src/Components/Focus.js
@@ -1,9 +1,10 @@
 import { useState } from 'react'
+
 import {
 	RiFocus2Line,
 	IcTwotoneCheckCircle,
 	OouiTrash,
-} from './../Icons/Icons'
+} from '../Icons'
 
 export const Focus = () => {
 	const [focus, setFocus] = useState(() =>
@@ -12,17 +13,18 @@ export const Focus = () => {
 
 	const [readOnly, setReadOnly] = useState(Boolean(focus))
 
-	const addToStorage = () => {
-		if (focus.length < 1) return
-		localStorage.setItem('focusToday', focus)
-		setReadOnly(true)
+	const toggleStorage = () => {
+		if (readOnly) {
+			localStorage.setItem('focusToday', '')
+			setFocus('')
+			setReadOnly(false)
+		} else {
+			if (focus.length < 1) return
+			localStorage.setItem('focusToday', focus)
+			setReadOnly(true)
+		}
 	}
 
-	const removeFromStorage = () => {
-		localStorage.setItem('focusToday', '')
-		setFocus('')
-		setReadOnly(false)
-	}
 	return (
 		<section className='flex'>
 			<div className='bg-inherit w-11/12 h-fit mb-2 mr-2 border-2 border-solid border-white relative'>
@@ -41,34 +43,11 @@ export const Focus = () => {
 					<p className='p-1 text-white w-full bg-inherit'>{focus}</p>
 				)}
 
-				{/* Make a single BTN IF POSSIBLE,  */}
-				{/* Problem: e.target.value doesnt work (no text inside btn) and there are 2 seperate func */}
-
-				{!readOnly ? (
-					<button
-						className='absolute left-[93%] top-[0%]'
-						value='delete-focus'
-						onClick={addToStorage}>
-						<IcTwotoneCheckCircle
-							className='inline-block'
-							width='1.5rem'
-							height='1.5rem'
-							pathfill='dark'
-						/>
-					</button>
-				) : (
-					<button
-						className='absolute left-[93%] top-[0%]'
-						value='delete-focus'
-						onClick={removeFromStorage}>
-						<OouiTrash
-							className='inline-block'
-							width='1.5rem'
-							height='1.5rem'
-							pathfill='white'
-						/>
-					</button>
-				)}
+				<button
+					className='block absolute left-[93%] top-[10%]'
+					onClick={toggleStorage}>
+					{!readOnly ? <IcTwotoneCheckCircle /> : <OouiTrash />}
+				</button>
 			</div>
 
 			<RiFocus2Line
